:sparkles: Allow customizing hotkey to open overlay
