+++ package.json
@@ -24,7 +24,8 @@
     "react-router": "^7.3.0",
     "react-syntax-highlighter": "^15.6.1",
     "remark-gfm": "^4.0.1",
-    "tailwindcss": "^4.0.13"
+    "tailwindcss": "^4.0.13",
+    "zustand": "^5.0.3"
   },
   "devDependencies": {
     "@tauri-apps/cli": "^2",
