+++ src/components/Overlay/Overlay.tsx
@@ -1,9 +1,12 @@
+import { useConfig } from '@/hooks/config';
 import { getCurrentWindow } from '@tauri-apps/api/window';
-import { register, unregister } from '@tauri-apps/plugin-global-shortcut';
+import { isRegistered, register, unregister } from '@tauri-apps/plugin-global-shortcut';
 import { useEffect, useState } from 'react';
 import { Outlet } from 'react-router';
 
 export function Overlay() {
+  const { config } = useConfig();
+
   const [canClose, setCanClose] = useState(false);
 
   useEffect(() => {
@@ -20,22 +23,6 @@ export function Overlay() {
       }
     });
 
-    register('Alt+Space', async e => {
-      if (e.state === 'Released') {
-        const isMinimized = await currentWindow.isMinimized();
-        const isVisible = await currentWindow.isVisible();
-
-        if (isMinimized || !isVisible) {
-          await currentWindow.unminimize();
-          await currentWindow.show();
-          await currentWindow.setFocus();
-        } else {
-          await currentWindow.hide();
-          await currentWindow.minimize();
-        }
-      }
-    });
-
     const escapeHandler = (e: KeyboardEvent) => {
       if (e.key === 'Escape') {
         currentWindow.hide();
@@ -47,10 +34,41 @@ export function Overlay() {
 
     return () => {
       unlisten.then(fn => fn());
-      unregister('Alt+Space');
       window.removeEventListener('keydown', escapeHandler);
     };
   }, []);
+
+  useEffect(() => {
+    async function init() {
+      const currentWindow = getCurrentWindow();
+
+      if (await isRegistered(config.hotkey || 'Alt+Space')) {
+        await unregister(config.hotkey || 'Alt+Space');
+      }
+
+      await register(config.hotkey || 'Alt+Space', async e => {
+        if (e.state === 'Released') {
+          const isMinimized = await currentWindow.isMinimized();
+          const isVisible = await currentWindow.isVisible();
+
+          if (isMinimized || !isVisible) {
+            await currentWindow.unminimize();
+            await currentWindow.show();
+            await currentWindow.setFocus();
+          } else {
+            await currentWindow.hide();
+            await currentWindow.minimize();
+          }
+        }
+      });
+    }
+
+    init();
+
+    return () => {
+      unregister(config.hotkey || 'Alt+Space');
+    };
+  }, [config.hotkey]);
 
   return (
     <div
